Creating ui file.
