+++ graphics/main.cpp
@@ -1,6 +1,6 @@
 // Copyright (c) 2016 Quenio Cesar Machado dos Santos. All rights reserved.
 
-#include "graphics.h"
+#include "ui.h"
 
 int main(int argc, char *argv[])
 {
